# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Kaay entreprendre",
        "description": "Cet api va nous permettre de faire communiquer la partie front et la partie back de notre plateforme",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/guide/create": {
            "post": {
                "tags": [
                    "Guides"
                ],
                "summary": "Ajouter un nouveau guide",
                "description": "Ajoute un nouveau guide avec les détails fournis",
                "operationId": "createGuide",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "phases": {
                                        "type": "string"
                                    },
                                    "reaction": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guide créé avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/guide/update/{id}": {
            "post": {
                "tags": [
                    "Guides"
                ],
                "summary": "Mettre à jour un guide existant",
                "description": "Met à jour les détails d'un guide existant en fonction de l'ID fourni",
                "operationId": "updateGuide",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du guide à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "phases": {
                                        "type": "string"
                                    },
                                    "reaction": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guide mis à jour avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/guide/archiver_guide/{id}": {
            "post": {
                "tags": [
                    "Guides"
                ],
                "summary": "Archiver un guide",
                "description": "Archive un guide en fonction de l'ID fourni",
                "operationId": "archiveGuide",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du guide à archiver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guide archivé avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/reponses": {
            "get": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Récupérer la liste des réponses",
                "description": "Récupère la liste de toutes les réponses",
                "operationId": "getReponses",
                "responses": {
                    "200": {
                        "description": "Liste des réponses récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/reponse/create": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Ajouter une nouvelle réponse",
                "description": "Ajoute une nouvelle réponse avec les détails fournis",
                "operationId": "createReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "commentaire_id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse ajoutée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/reponse/edit": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Mettre à jour une réponse existante",
                "description": "Met à jour les détails d'une réponse existante",
                "operationId": "updateReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse mise à jour avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/reponse/archive": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Archiver une réponse",
                "description": "Archive une réponse en fonction de l'ID fourni",
                "operationId": "archiveReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse archivée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/reponse/delete": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Supprimer une réponse",
                "description": "Supprime une réponse en fonction de l'ID fourni",
                "operationId": "destroyReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}